refactor: checkbox types
